src/chatbot/pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["rieh01 <hoxharie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
sentence-transformers = "^3.0.1"
scikit-learn = "^1.5.0"
numpy = "<2.0"
torch = "^2.3.1"
transformers = "^4.41.2"
flask = "^3.0.3"
google-auth = "^2.30.0"
google-cloud-dialogflow = "^2.30.0"
huggingface-hub = "^0.23.4"
sympy = "^1.12.1"
accelerate = "^0.31.0"
python-dotenv = "^1.0.1"
ngrok = "^1.3.0"
sentencepiece = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
